feat(web): run migration in place from the review screen
Merge the separate Importing step into the review screen so the Import button
stays put and shows a loading state while the migration runs, instead of the
button vanishing on confirm. Per-item cards update live and a truthful summary
plus "Open the new class" replace the button on completion. Remove the now-
unused ExecuteStep and its dead strings.
